Split lessons routes to avoid route-entry freeze
