+++ client/src/App.tsx
@@ -16,8 +16,8 @@ const LocalGame = lazy(loadLocalGameRoute);
 const QuickPlay = lazy(loadQuickPlayRoute);
 const BotGame = lazy(loadBotGameRoute);
 const PuzzleStreakPage = lazy(() => import('./routes/PuzzleRoutes').then(m => ({ default: m.PuzzleStreakRoute })));
-const LessonCoursePage = lazy(() => import('./routes/LessonsRoutes').then(m => ({ default: m.LessonCourseRoute })));
-const LessonPlayerPage = lazy(() => import('./routes/LessonsRoutes').then(m => ({ default: m.LessonPlayerRoute })));
+const LessonCoursePage = lazy(() => import('./routes/LessonCourseRoute'));
+const LessonPlayerPage = lazy(() => import('./routes/LessonPlayerRoute'));
 const PuzzlePlayer = lazy(() => import('./routes/PuzzleRoutes').then(m => ({ default: m.PuzzlePlayerRoute })));
 const AboutPage = lazy(() => import('./components/AboutPage'));
 const GamesPage = lazy(() => import('./components/GamesPage'));
