+++ client/src/routes/LessonCourseRoute.tsx
@@ -1,0 +1,10 @@
+import { LessonProgressProvider } from '../lib/lessonProgress';
+import LessonCoursePage from '../components/LessonCoursePage';
+
+export default function LessonCourseRoute() {
+  return (
+    <LessonProgressProvider>
+      <LessonCoursePage />
+    </LessonProgressProvider>
+  );
+}
