+++ client/src/routes/LessonPlayerRoute.tsx
@@ -1,0 +1,10 @@
+import { LessonProgressProvider } from '../lib/lessonProgress';
+import LessonPlayerPage from '../components/LessonPlayerPage';
+
+export default function LessonPlayerRoute() {
+  return (
+    <LessonProgressProvider>
+      <LessonPlayerPage />
+    </LessonProgressProvider>
+  );
+}
