--- client/src/routes/LessonsRoutes.tsx
@@ -1,18 +1,0 @@
-import { LessonProgressProvider } from '../lib/lessonProgress';
-import { LessonCoursePage, LessonPlayerPage } from '../components/LessonsPage';
-
-export function LessonCourseRoute() {
-  return (
-    <LessonProgressProvider>
-      <LessonCoursePage />
-    </LessonProgressProvider>
-  );
-}
-
-export function LessonPlayerRoute() {
-  return (
-    <LessonProgressProvider>
-      <LessonPlayerPage />
-    </LessonProgressProvider>
-  );
-}
